+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.2.0
+* Add instant iterative development mode. `sparkly-testing --help` for more details.
+* Use in-memory db for Hive Metastore in `SparklyTest` (faster tests).
+* `spark.sql.shuffle.partitions = 4` for `SparklyTest` (faster tests).
+* `spark.sql.warehouse.dir = <random tmp dir>` for `SparklyTest` (no side effects)
+
 ## 2.1.1
 * Fix: remove backtick quoting from catalog utils to ease work with different databases.